requirements.txt
Django
djangorestframework
psycopg2
python-decouple
flake8